fixed the fucking billion errors
every time the site would regenerate, the command prompt would scream the same error over and over again. this happens only when you are using a liquid variable within a file path within a liquid modifier, and for some reason, the way to fix it is to put the file path with the variable inside of a capture and use whatever you named the capture instead of just putting the filepath in. this makes absolutely no sense, as nothing is technically changing, but if i had to guess why this happens, it'd probably be because different syntax's for different lnguages* are clashing and trying to do different things. I honestly have no idea though, that's just a guess. I am an insane man